Rename ironic go module to correct path for go get can find
`go get github.com/metal3-io/baremetal-operator/ironic` fails because
it can  not find new ironic module which lies under the pkg directory.
This PR fixes this bug by renaming ironic module to be parallel with folder
it lies.

LICENSE file is also added for GOPROXY url
can cache these modules in pkg.go.dev.
