+++ pkg/ironic/raid.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/gophercloud/gophercloud/openstack/baremetal/v1/nodes"
 	metal3v1alpha1 "github.com/metal3-io/baremetal-operator/apis/metal3.io/v1alpha1"
-	"github.com/metal3-io/baremetal-operator/ironic/devicehints"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/devicehints"
 )
 
 const (
