+++ pkg/provisioner/demo/demo.go
@@ -3,7 +3,7 @@ package demo
 import (
 	"time"
 
-	"github.com/metal3-io/baremetal-operator/ironic/bmc"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/bmc"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
