+++ pkg/provisioner/ironic/ironic.go
@@ -5,11 +5,11 @@ import (
 	"strings"
 	"time"
 
-	"github.com/metal3-io/baremetal-operator/ironic"
-	"github.com/metal3-io/baremetal-operator/ironic/devicehints"
-	"github.com/metal3-io/baremetal-operator/ironic/hardwaredetails"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/devicehints"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/hardwaredetails"
 
-	"github.com/metal3-io/baremetal-operator/ironic/bmc"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/bmc"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
