+++ pkg/provisioner/ironic/ironic_test.go
@@ -3,8 +3,8 @@ package ironic
 import (
 	"testing"
 
-	"github.com/metal3-io/baremetal-operator/ironic/bmc"
-	"github.com/metal3-io/baremetal-operator/ironic/clients"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/bmc"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/clients"
 
 	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -17,7 +17,7 @@ import (
 	// We don't use this package directly here, but need it imported
 	// so it registers its test fixture with the other BMC access
 	// types.
-	_ "github.com/metal3-io/baremetal-operator/ironic/testbmc"
+	_ "github.com/metal3-io/baremetal-operator/pkg/ironic/testbmc"
 )
 
 func init() {
