+++ pkg/provisioner/ironic/provisioncapacity_test.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/metal3-io/baremetal-operator/ironic/bmc"
-	"github.com/metal3-io/baremetal-operator/ironic/clients"
-	"github.com/metal3-io/baremetal-operator/ironic/testserver"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/bmc"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/clients"
+	"github.com/metal3-io/baremetal-operator/pkg/ironic/testserver"
 
 	"github.com/gophercloud/gophercloud/openstack/baremetal/v1/nodes"
 	"github.com/stretchr/testify/assert"
